+++ lib/app_setup.dart
@@ -6,6 +6,7 @@
 // ignore_for_file: prefer-match-file-name
 
 import 'dart:convert';
+import 'dart:io';
 
 import 'package:common/data/adapters/uri_adapter.dart';
 import 'package:common/data/model/hive/gcode_macro.dart';
@@ -16,6 +17,7 @@ import 'package:common/data/model/hive/octoeverywhere.dart';
 import 'package:common/data/model/hive/progress_notification_mode.dart';
 import 'package:common/data/model/hive/remote_interface.dart';
 import 'package:common/data/model/hive/temperature_preset.dart';
+import 'package:common/exceptions/mobileraker_exception.dart';
 import 'package:common/service/firebase/analytics.dart';
 import 'package:common/service/firebase/auth.dart';
 import 'package:common/service/firebase/remote_config.dart';
@@ -103,13 +105,24 @@ setupBoxes() async {
       element.save();
     });
   } catch (e) {
-    logger.e(
-      'There was an error while trying to init Hive. Resetting all Hive data...',
-    );
-    await Hive.deleteBoxFromDisk('printers');
-    await Hive.deleteBoxFromDisk('uuidbox');
-    await Hive.deleteBoxFromDisk('settingsbox');
-    await openBoxes(keyMaterial);
+    if (e is TypeError) {
+      logger.e('An TypeError occurred while trying to open Boxes...', e);
+      logger.e('Will reset all stored data to resolve this issue!');
+      throw MobilerakerStartupException(
+        'An unexpected TypeError occurred while parsing the stored app data. Please report this error to the developer. To resolve this issue clear the app storage or reinstall the app.',
+        parentException: e,
+        canResetStorage: true,
+      );
+    } else if (e is FileSystemException) {
+      logger.e('An FileSystemException(${e.runtimeType}) occured while trying to open Boxes...', e);
+      throw MobilerakerStartupException(
+        'Failed to retrieve app data from system storage. Please restart the app. If the error persists, consider clearing the storage or reinstalling the app.',
+        parentException: e,
+        canResetStorage: true,
+      );
+    }
+    logger.e('An unexpected error occurred while trying to open Boxes...', e);
+    rethrow;
   }
   logger.i('Completed Hive init');
 }
@@ -156,6 +169,17 @@ Future<List<Box>> openBoxes(Uint8List _) {
     Hive.openBox('settingsbox'),
     Hive.openBox<Notification>('notifications'),
     // Hive.openBox<OctoEverywhere>('octo', encryptionCipher: HiveAesCipher(keyMaterial))
+  ]);
+}
+
+Future<void> deleteBoxes() {
+  logger.i('Deleting all boxes');
+  return Future.wait([
+    Hive.deleteBoxFromDisk('printers'),
+    Hive.deleteBoxFromDisk('uuidbox'),
+    Hive.deleteBoxFromDisk('settingsbox'),
+    Hive.deleteBoxFromDisk('notifications'),
+    // Hive.deleteBoxFromDisk('octo')
   ]);
 }