+++ lib/ui/components/dialog/webcam_preview_dialog.dart
@@ -35,6 +35,7 @@ class WebcamPreviewDialog extends HookWidget {
     WebcamPreviewDialogArguments arg = request.data;
 
     return Dialog(
+      clipBehavior: Clip.antiAlias,
       child: Padding(
         padding: const EdgeInsets.all(4.0),
         child: ConstrainedBox(