+++ lib/ui/screens/dashboard/dashboard_page.dart
@@ -170,7 +170,7 @@ class _BottomNavigationBar extends HookConsumerWidget {
     var themeData = Theme.of(context);
     var colorScheme = themeData.colorScheme;
 
-    if (ref.watch(machinePrinterKlippySettingsProvider.select((value) => value.isReloading || value.hasError))) {
+    if (ref.watch(machinePrinterKlippySettingsProvider.select((value) => value.isLoading && !value.isReloading))) {
       return const SizedBox.shrink();
     }
 
@@ -186,7 +186,11 @@ class _BottomNavigationBar extends HookConsumerWidget {
       notchSmoothness: NotchSmoothness.softEdge,
       activeIndex: activeIndex,
       splashSpeedInMilliseconds: kThemeAnimationDuration.inMilliseconds,
-      onTap: (index) => pageController.animateToPage(index, duration: kThemeChangeDuration, curve: Curves.easeOutCubic),
+      onTap: (index) {
+        if (pageController.hasClients) {
+          pageController.animateToPage(index, duration: kThemeChangeDuration, curve: Curves.easeOutCubic);
+        }
+      },
     );
   }
 }