+++ docs/changelog.md
@@ -9,6 +9,8 @@
 ### Bug Fixes
 
 - [IOS] Fixed live activities foreground color on IOS 16.4 [#315](https://github.com/Clon1998/mobileraker/issues/315)
+- [IOS] Fixed an issue with the live activities that would reset the entire local storage of the
+  app [#315](https://github.com/Clon1998/mobileraker/issues/321)
 - Fixed typo on profile page's 'Send Verification Mail' button
 
 ## [2.6.9] - 2024-01-10